jenkins/requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    make upgrade
#
asgiref==3.4.1
    # via django
astroid==2.9.0
    # via
    #   -r requirements/quality.txt
    #   pylint
    #   pylint-celery
attrs==21.2.0
    # via
    #   -r requirements/quality.txt
    #   pytest
backports.entry-points-selectable==1.1.1
    # via virtualenv
boto==2.49.0
    # via -r requirements/quality.txt
boto3==1.20.26
    # via -r requirements/quality.txt
botocore==1.23.26
    # via
    #   -r requirements/quality.txt
    #   boto3
    #   s3transfer
certifi==2021.10.8
    # via
    #   -r requirements/quality.txt
    #   -r requirements/travis.txt
    #   requests
cffi==1.15.0
    # via
    #   -r requirements/quality.txt
    #   cryptography
chardet==4.0.0
    # via diff-cover
charset-normalizer==2.0.9
    # via
    #   -r requirements/quality.txt
    #   -r requirements/travis.txt
    #   requests
click==8.0.3
    # via
    #   -r requirements/pip-tools.txt
    #   -r requirements/quality.txt
    #   click-log
    #   code-annotations
    #   edx-lint
    #   pip-tools
click-log==0.3.2
    # via
    #   -r requirements/quality.txt
    #   edx-lint
code-annotations==1.2.0
    # via
    #   -r requirements/quality.txt
    #   edx-lint
codecov==2.1.12
    # via -r requirements/travis.txt
coverage[toml]==6.2
    # via
    #   -r requirements/quality.txt
    #   -r requirements/travis.txt
    #   codecov
    #   pytest-cov
cryptography==3.4.8
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/quality.txt
    #   pyopenssl
diff-cover==6.4.4
    # via -r requirements/dev.in
distlib==0.3.4
    # via virtualenv
django==3.2.10
    # via edx-i18n-tools
edx-i18n-tools==0.8.1
    # via -r requirements/dev.in
edx-lint==5.2.1
    # via -r requirements/quality.txt
filelock==3.4.0
    # via
    #   tox
    #   virtualenv
idna==3.3
    # via
    #   -r requirements/quality.txt
    #   -r requirements/travis.txt
    #   requests
iniconfig==1.1.1
    # via
    #   -r requirements/quality.txt
    #   pytest
isort==5.10.1
    # via
    #   -r requirements/quality.txt
    #   pylint
jinja2==3.0.3
    # via
    #   -r requirements/quality.txt
    #   code-annotations
    #   diff-cover
jmespath==0.10.0
    # via
    #   -r requirements/quality.txt
    #   boto3
    #   botocore
lazy-object-proxy==1.7.1
    # via
    #   -r requirements/quality.txt
    #   astroid
markupsafe==2.0.1
    # via
    #   -r requirements/quality.txt
    #   jinja2
mccabe==0.6.1
    # via
    #   -r requirements/quality.txt
    #   pylint
packaging==21.3
    # via
    #   -r requirements/quality.txt
    #   pytest
    #   tox
path==16.2.0
    # via edx-i18n-tools
pbr==5.8.0
    # via
    #   -r requirements/quality.txt
    #   stevedore
pep517==0.12.0
    # via
    #   -r requirements/pip-tools.txt
    #   pip-tools
pip-tools==6.4.0
    # via -r requirements/pip-tools.txt
platformdirs==2.4.0
    # via
    #   -r requirements/quality.txt
    #   pylint
    #   virtualenv
pluggy==1.0.0
    # via
    #   -r requirements/quality.txt
    #   diff-cover
    #   pytest
    #   tox
polib==1.1.1
    # via edx-i18n-tools
py==1.11.0
    # via
    #   -r requirements/quality.txt
    #   pytest
    #   tox
pycodestyle==2.8.0
    # via -r requirements/quality.txt
pycparser==2.21
    # via
    #   -r requirements/quality.txt
    #   cffi
pydocstyle==6.1.1
    # via -r requirements/quality.txt
pygments==2.10.0
    # via diff-cover
pylint==2.12.2
    # via
    #   -r requirements/quality.txt
    #   edx-lint
    #   pylint-celery
    #   pylint-django
    #   pylint-plugin-utils
pylint-celery==0.3
    # via
    #   -r requirements/quality.txt
    #   edx-lint
pylint-django==2.4.4
    # via
    #   -r requirements/quality.txt
    #   edx-lint
pylint-plugin-utils==0.6
    # via
    #   -r requirements/quality.txt
    #   pylint-celery
    #   pylint-django
pymongo==4.0.1
    # via -r requirements/quality.txt
pymysql==1.0.2
    # via -r requirements/quality.txt
pyopenssl==21.0.0
    # via -r requirements/quality.txt
pyparsing==3.0.6
    # via
    #   -r requirements/quality.txt
    #   packaging
pytest==6.2.5
    # via
    #   -r requirements/quality.txt
    #   pytest-cov
    #   pytest-django
pytest-cov==3.0.0
    # via -r requirements/quality.txt
pytest-django==4.5.2
    # via -r requirements/quality.txt
python-dateutil==2.8.2
    # via
    #   -r requirements/quality.txt
    #   botocore
python-gnupg==0.4.8
    # via -r requirements/quality.txt
python-slugify==5.0.2
    # via
    #   -r requirements/quality.txt
    #   code-annotations
pytz==2021.3
    # via django
pyyaml==6.0
    # via
    #   -r requirements/quality.txt
    #   code-annotations
    #   edx-i18n-tools
requests==2.26.0
    # via
    #   -r requirements/quality.txt
    #   -r requirements/travis.txt
    #   codecov
s3transfer==0.5.0
    # via
    #   -r requirements/quality.txt
    #   boto3
six==1.16.0
    # via
    #   -r requirements/quality.txt
    #   edx-lint
    #   pyopenssl
    #   python-dateutil
    #   tox
    #   virtualenv
snowballstemmer==2.2.0
    # via
    #   -r requirements/quality.txt
    #   pydocstyle
sqlparse==0.4.2
    # via django
stevedore==3.5.0
    # via
    #   -r requirements/quality.txt
    #   code-annotations
text-unidecode==1.3
    # via
    #   -r requirements/quality.txt
    #   python-slugify
toml==0.10.2
    # via
    #   -r requirements/quality.txt
    #   pylint
    #   pytest
    #   tox
tomli==2.0.0
    # via
    #   -r requirements/pip-tools.txt
    #   -r requirements/quality.txt
    #   coverage
    #   pep517
tox==3.24.4
    # via tox-battery
tox-battery==0.6.1
    # via -r requirements/dev.in
typing-extensions==4.0.1
    # via
    #   -r requirements/quality.txt
    #   astroid
    #   pylint
urllib3==1.26.7
    # via
    #   -r requirements/quality.txt
    #   -r requirements/travis.txt
    #   botocore
    #   requests
virtualenv==20.10.0
    # via tox
wheel==0.37.1
    # via
    #   -r requirements/pip-tools.txt
    #   pip-tools
wrapt==1.13.3
    # via
    #   -r requirements/quality.txt
    #   astroid

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
